centralize common types
